Added Dockerfile to run the frontend project with React in a Docker container
Added a Dockerfile that specifies the creation of a Docker image and running the frontend project with React in a Docker container. This includes installing dependencies and starting the React server with the npm start command, allowing the project to start and run smoothly in a Docker container without issues. This makes it possible to easily deploy the Docker container to a Production environment.